Added error messages stating values cannot be null on AddPartController.java, AddInhousePartController.java,AddOutsourcedPartController.java, and Part.java. Added input fields are required on InhousePartForm.html, OutsourcedPartForm.html, and productForm.html. Updated README.md.

+++ src/main/java/com/example/demo/controllers/AddPartController.java
@@ -56,6 +56,10 @@ public class AddPartController {
     public String deletePart(@Valid @RequestParam("partID") int theId,  Model theModel){
         PartService repo = context.getBean(PartServiceImpl.class);
         Part part=repo.findById(theId);
+
+        if (part == null) {
+            return "errorpartnotfound";
+        }
         if(part.getProducts().isEmpty()){
             repo.deleteById(theId);
             return "confirmationdeletepart";
